refactor: refactor DID module types and move to controller package
