+++ x/did/keeper/controller.go
@@ -3,7 +3,8 @@ package keeper
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/onsonr/crypto/mpc"
-	"github.com/onsonr/sonr/x/did/types"
+
+	types "github.com/onsonr/sonr/x/did/types/controller"
 )
 
 func (k Keeper) NewController(ctx sdk.Context) (types.ControllerI, error) {
