--- x/did/types/account.go
@@ -1,12 +1,0 @@
-package types
-
-type SonrAccount struct {
-	ID         string
-	Name       string
-	Address    string
-	PublicKey  string
-	ChainCode  string
-	Index      string
-	Controller string
-	CreatedAt  string
-}
