+++ x/did/types/controller/address.go
@@ -1,4 +1,4 @@
-package types
+package controller
 
 import (
 	"crypto/ecdsa"
