+++ x/did/types/controller/controller.go
@@ -1,9 +1,9 @@
-package types
+package controller
 
 import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/onsonr/crypto/mpc"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	didv1 "github.com/onsonr/sonr/api/did/v1"
 )
