+++ x/did/types/controller/secpk251.go
@@ -1,4 +1,4 @@
-package mpc
+package controller
 
 import (
 	"bytes"
