+++ x/did/types/controller/signer.go
@@ -1,4 +1,4 @@
-package mpc
+package controller
 
 import (
 	"context"
