+++ x/vault/keeper/querier.go
@@ -7,8 +7,8 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/onsonr/crypto/mpc"
-	didtypes "github.com/onsonr/sonr/x/did/types"
 
+	"github.com/onsonr/sonr/x/did/types/controller"
 	"github.com/onsonr/sonr/x/vault/types"
 )
 
@@ -72,7 +72,7 @@ func (k Querier) Allocate(goCtx context.Context, req *types.QueryAllocateRequest
 	}
 
 	// 3. Create Controller from Keyshares
-	con, err := didtypes.NewController(shares)
+	con, err := controller.NewController(shares)
 	if err != nil {
 		ctx.Logger().Error(fmt.Sprintf("Error creating controller: %s", err.Error()))
 		return nil, types.ErrControllerCreation.Wrap(err.Error())
